backend/requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple
aerich==0.9.2
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiomysql==0.2.0
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
Appium-Python-Client==2.11.1
appnope==0.1.4
APScheduler==3.10.4
asttokens==3.0.1
async-timeout==5.0.1
asyncclick==8.3.0.7
asyncpg==0.29.0
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2023.7.22
cffi==1.17.0
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
crontab==1.0.1
cryptography==41.0.3
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
distro==1.9.0
docopt==0.6.2
docx2txt==0.9
emoji==2.15.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==2.2.1
Faker==8.10.1
fastapi==0.128.0
fastjsonschema==2.21.2
filetype==1.2.0
frozenlist==1.8.0
gevent==24.2.1
greenlet==3.3.0
grpcio==1.76.0
gunicorn==21.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.4
iniconfig==2.3.0
ipython==8.12.3
iso8601==2.1.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-mcp-adapters==0.2.1
langchain-openai==1.1.6
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
loguru==0.7.0
lxml==6.0.2
Markdown==3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mcp==1.25.0
mistune==3.2.0
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.7
nltk==3.9.2
numpy==1.26.4
olefile==0.47
openai==2.14.0
openpyxl==3.0.9
orjson==3.11.5
ormsgpack==1.12.1
outcome==1.2.0
packaging==23.2
pactverify==2.0.1
pandas==3.0.0
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.5
pdfminer==20191125
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pickleshare==0.7.5
pika==1.3.2
pillow==12.1.0
pipreqs==0.5.0
platformdirs==4.5.1
playwright==1.57.0
pluggy==1.6.0
portalocker==2.7.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.21
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pypdf==6.6.0
PyPDF2==3.0.1
pypdfium2==5.3.0
pypika-tortoise==0.6.3
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.21
python-oxmsg==0.0.2
python-pptx==1.0.2
python3-saml==1.15.0
pytz==2023.3
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.16.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
selenium==4.1.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.45
sqlite-vec==0.1.6
sse-starlette==3.0.4
stack-data==0.6.3
starlette==0.49.1
stomp-py==8.1.2
tenacity==8.5.0
text-unidecode==1.3
tiktoken==0.12.0
tinycss2==1.4.0
tomlkit==0.11.8
tornado==6.5.4
tortoise-orm==0.25.3
tqdm==4.67.1
traitlets==5.14.3
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
tzlocal==5.0.1
unstructured==0.18.27
unstructured-client==0.42.8
urllib3==1.26.16
urllib3-secure-extra==0.1.0
uuid_utils==0.12.0
uvicorn==0.34.3
validators==0.20.0
wcwidth==0.5.0
webencodings==0.5.1
websocket-client==1.8.0
win32_setctime==1.2.0
wrapt==2.0.1
wsproto==1.2.0
xlrd==2.0.1
xlsxwriter==3.2.9
XMind==1.2.0
xmlsec==1.3.17
xxhash==3.6.0
yarg==0.1.9
yarl==1.22.0
zope.event==5.0
zope.interface==6.0
zstandard==0.23.0
